Total Update
The framework was updated significantly to be ready as a version
release.

The system now uses Python 2.7.9, to be available to the large amount of
other currently available Python libraries for that version. With some
extra work, it should be able to run on Python 3 for future adaptation.

A lot of the code was cleaned up to be more readable and modular. Each
file was edited using the PyScripter IDE to keep the whitespace
organized.

A lot of the DIR_ and DAT_ name conventions were taken out. Console was
renamed Framesys (for Frame System). Now, things like sprites and sound
data are designed to only load once so each subsequent instance doesn't
have to load extra times.