More Minor Changes
Made it so every sprite in every group will receive a DF change, so all
motion-related activity stays in the sprites. DF was also tabbed in the
wrong spot, so that was adjusted as well to compliment the above change.

Sprite Groups are now compiled into one easy 'for' loop.